refactor: move context file to proper folder with proper name
