+++ src/components/period-picker-with-period-type/PeriodPickerWithPeriodType.js
@@ -8,7 +8,7 @@ import i18n from '@dhis2/d2-i18n';
 import { DropDown, PeriodPicker } from '@dhis2/d2-ui-core';
 
 /* App context */
-import AppContext from '../../context';
+import AppContext from '../../pages/AppContext';
 
 /* Actions */
 import loadPeriodTypes from '../../redux/actions/periodTypes';
