+++ src/pages/data-approval/DataApprovalStatus.js
@@ -9,7 +9,7 @@ import { Dialog } from 'material-ui';
 import { Button } from '@dhis2/d2-ui-core';
 
 /* App context */
-import AppContext from '../../context';
+import AppContext from '../../pages/AppContext';
 
 /* i18n */
 import i18n from '../../locales';
